Move query params command API test to new file
