+++ lib/tests/streamlit/commands/query_params_test.py
@@ -1,0 +1,41 @@
+# Copyright (c) Streamlit Inc. (2018-2022) Snowflake Inc. (2022)
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import streamlit as st
+from streamlit.errors import StreamlitAPIException
+from tests.delta_generator_test_case import DeltaGeneratorTestCase
+
+
+class QueryParamsAPITest(DeltaGeneratorTestCase):
+    """Test Query params commands APIs."""
+
+    def test_set_query_params_sends_protobuf_message(self):
+        """Test valid st.set_query_params sends protobuf message."""
+        st.experimental_set_query_params(x="a")
+        message = self.get_message_from_queue(0)
+        self.assertEqual(message.page_info_changed.query_string, "x=a")
+
+    def test_set_query_params_exceptions(self):
+        """Test invalid st.set_query_params raises exceptions."""
+        with self.assertRaises(StreamlitAPIException):
+            st.experimental_set_query_params(embed="True")
+        with self.assertRaises(StreamlitAPIException):
+            st.experimental_set_query_params(embed_options="show_colored_line")
+
+    def test_get_query_params_after_set_query_params(self):
+        """Test valid st.set_query_params sends protobuf message."""
+        p_set = dict(x=["a"])
+        st.experimental_set_query_params(**p_set)
+        p_get = st.experimental_get_query_params()
+        self.assertEqual(p_get, p_set)
