+++ lib/tests/streamlit/streamlit_test.py
@@ -179,23 +179,3 @@ class StreamlitAPITest(DeltaGeneratorTestCase):
     def test_st_legacy_vega_lite_chart(self):
         """Test st._legacy_vega_lite_chart."""
         pass
-
-    def test_set_query_params_sends_protobuf_message(self):
-        """Test valid st.set_query_params sends protobuf message."""
-        st.experimental_set_query_params(x="a")
-        message = self.get_message_from_queue(0)
-        self.assertEqual(message.page_info_changed.query_string, "x=a")
-
-    def test_set_query_params_exceptions(self):
-        """Test invalid st.set_query_params raises exceptions."""
-        with self.assertRaises(StreamlitAPIException):
-            st.experimental_set_query_params(embed="True")
-        with self.assertRaises(StreamlitAPIException):
-            st.experimental_set_query_params(embed_options="show_colored_line")
-
-    def test_get_query_params_after_set_query_params(self):
-        """Test valid st.set_query_params sends protobuf message."""
-        p_set = dict(x=["a"])
-        st.experimental_set_query_params(**p_set)
-        p_get = st.experimental_get_query_params()
-        self.assertEqual(p_get, p_set)
